pyproject.toml
[project]
name = "joseph"
version = "0.1.0"
description = "Hybrid AI text detection system"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "transformers>=4.47.0",
    "torch>=2.9.0",
    "numpy>=1.26.0,<3.0",
    "pydantic>=2.10.0",
    "pydantic-settings>=2.6.0",
    "python-multipart>=0.0.20",
    "jinja2>=3.1.0",
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "python-jose[cryptography]>=3.3.0",
    "httpx>=0.28.0",
    "joblib>=1.4.0",
    "scikit-learn>=1.5.0",
]

[project.optional-dependencies]
training = [
    "pandas>=2.2.0",
    "datasets>=3.2.0",
    "tqdm>=4.67.0",
    "pyarrow>=18.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.bandit]
exclude_dirs = ["/test", "tests"]
skips = []

[tool.black]
line-length = 100
target-version = ['py311']

[tool.isort]
profile = "black"
line_length = 100
